# Add Initial address canonization support