--- a/volatility3/framework/layers/linear.py
+++ b/volatility3/framework/layers/linear.py
@@ -79,3 +79,4 @@ class LinearlyMappedLayer(interfaces.layers.TranslationLayerInterface):
     def get_canonical_address(cls, address: int) -> int:
         """Make the address canonical"""
         return address
+