This PR fixes #699.
Looking forward for more testing and comments.