@Digitalisx 2022-04-24
@paulkermann Thank you for reflecting on my little review, although it would be nice to have in-depth reviews and comments on memory... But I think I have enough knowledge not yet. I always looking forward to your good contribution and hope that your thoughts will be conveyed to Maintainers well. 😊