@paulkermann 2022-04-24
With this PR accessing non canonical addresses will cause a python exception.
This can be rectified for backward compatibility if needed.