@ikelos 2022-04-24
@Digitalisx, we've discussed this with @paulkermann offline.  It's quite a significant change to the way volatility used to work, and the justification for making such a change hasn't really been made yet.  Accessing non-canonical addresses would need to still work in case it's relied upon by other plugins.  That's one of the reasons I'm very reticent to make this change, it doesn't feel like the limited benefit outweighs the potential costs.