@Digitalisx 2022-04-24
@ikelos First of all, thank you for letting me know about the offline discussion related this issue!

That's right, I saw your concern mentioned in this PR related issue.
So I checked if there are any improvements or differences in existing codes for the code presented, apart from the actual merge of the requested PRs.
The core area processing with memory is a stage where I am learning by looking at their code, so I am also careful about giving opinions.

As in this case, I have been reviewing issues and PRs these days, or making some comments, and I am often worried that giving wrong information or direction. Please let me know if you think the timing of the review is not appropriate other than operational problems or functional problems.
(I don't want my passion to harm the community.. 😅)