@ikelos 2022-04-24
Hi @Digitalisx, reviewing PRs and so on is fine as long as you're clear about your experience levels and that you're not part of the core team.  We're happy to get community contributions, and happy to have the community review itself as long as everyone's got their eyes open as to what's going on.  If you make mistakes my hope is that you'll be (gently) corrected and the right information wins out...  5:)