@paulkermann 2022-04-24
This PR is now backward compatible. Read from non-canonical addresses will log a debug message when an invalid address is being read but will still continue.