@ikelos 2022-06-13
For completeness, I found another win10 x64 image and ran `psscan` which gave the following timings:

Without changes: 42.98s user 0.63s system 89% cpu 48.566 total
With changes: 91.39s user 9.34s system 12% cpu 13:38.77 total

That 18x times (an order of magnitude) slower because the address space maximum address has been increased.